extract BiostoreConnector interface
